DryTooling

Documentation for DryTooling.

DryTooling.Kramers

DryTooling.Kramers.RotaryKilnBedGeometry

DryTooling.Kramers.RotaryKilnBedGeometry

DryTooling.Kramers.SolutionLinearKramersModel

DryTooling.Kramers.SymbolicLinearKramersModel

DryTooling.Kramers.plotlinearkramersmodel

DryTooling.Kramers.solvelinearkramersmodel

DryTooling.Kramers — Module

module Kramers

Implements the differential equation for prediction of bed height profile in a rotary kiln as proposed by Kramers and Croockewite (1952) [1].

DryTooling.Kramers.RotaryKilnBedGeometry — Type

RotaryKilnBedGeometry

Description of a rotary kiln bed geometry computed from the solution of bed height along the kiln length. The main goal of the quantities computed here is their use with heat and mass transfer models for the simulation of rotary kiln process.

DryTooling.Kramers.RotaryKilnBedGeometry — Method

RotaryKilnBedGeometry(
    z::Vector{Float64},
    h::Vector{Float64},
    R::Float64,
    L::Float64
)

z::Vector{Float64}, solution coordinates over length, [m].

h::Vector{Float64}, bed profile solution over length, [m].

R::Float64, kiln internal radius, [m].

L::Float64, kiln length, [m].

DryTooling.Kramers.SolutionLinearKramersModel — Type

SolutionLinearKramersModel

Solve and process results from a symbolic linear Kramers model.

DryTooling.Kramers.SymbolicLinearKramersModel — Type

    SymbolicLinearKramersModel

Creates a reusable linear Kramers model for rotary kiln simulation.

DryTooling.Kramers.plotlinearkramersmodel — Method

plotlinearkramersmodel(
    model::SolutionLinearKramersModel;
    normz::Bool = false,
    normh::Bool = false
)::Any

Display plot of model solution for rotary kiln bed profile. Arguments normz and normh control whether z-coordinate and bed height must be normalized, respectively.

DryTooling.Kramers.solvelinearkramersmodel — Method

solvelinearkramersmodel(;
    L::Float64,
    D::Float64,
    Φ::Float64,
    ω::Float64,
    β::Float64,
    γ::Float64,
    d::Float64,
    model::Union{SymbolicLinearKramersModel,Nothing} = nothing
)::SolutionLinearKramersModel

L: kiln length, [m].

D: kiln internal diameter, [m].

Φ: kiln feed rate, [m³/h].

ω: kiln rotation rate, [rev/min].

β: kiln slope, [°].

γ: solids repose angle, [°].

d: particle size or dam height, [mm].

Important: inputs are in customary units based in international system (SI) - no imperial units supported - while all the computed values are provided in SI units as a better physical practice. If d represents a particle size, take care that it is provided in millimeters for ease of exchange with common dam sizes.

1https://doi.org/10.1016/0009-2509(52)87019-8