DryTooling.General

DryTooling.IdealGas.GAS_CONSTANT

DryTooling.Transport.MODELS

DryTooling.IdealGas.AbstractGasThermo

DryTooling.IdealGas.IdealGasMixture

DryTooling.IdealGas.IdealGasSpecies

DryTooling.IdealGas.IdealGasSpecies

DryTooling.IdealGas.IdealGasThermo

DryTooling.Kramers.RotaryKilnBedSolution

DryTooling.Kramers.RotaryKilnBedSolution

DryTooling.Kramers.SymbolicLinearKramersModel

DryTooling.Kramers.SymbolicLinearKramersModel

DryTooling.Transport.AbstractTransportModel

DryTooling.Transport.LennardJonesTransport

DryTooling.General.heaviside

DryTooling.General.interval

DryTooling.General.makestepwise1d

DryTooling.IdealGas.getthermo

DryTooling.IdealGas.nasa7enthapy

DryTooling.IdealGas.nasa7specificheat

DryTooling.Kramers.plotlinearkramersmodel

DryTooling.Kramers.solvelinearkramersmodel

DryTooling.General.heaviside — Method

Automatic differentiable Heaviside function.

DryTooling.General.interval — Method

Returns 1 if $x ∈ (a, b)$, 1/2 for $x = a || x = b$, or 0 .

DryTooling.General.makestepwise1d — Method

makestepwise1d(lo, hi, xc)

Creates an univariate function that is composed of two parts, the first evaluated before a critical domain point xc, and the seconda above that value. This is often required, for instance, for the evaluation of NASA polynomials for thermodynamic properties. If differentiable, then the returned function is compatible with symbolic argument as required when using package ModelingToolkit, etc.