DryTooling.FiniteVolumes

Grid generation

DryTooling.FiniteVolumes.UserDefinedGrid1D — Type

A minimal 1D grid with cells centers and walls.

Provides a very simple interface for the 1D grids used in standard finite volume solvers provided in the package. Constructor accepts a vector of coordinates and compute walls at mid-points between cells centers. First and last cells are over the boundaries and the models must consider this in the implementation.

N::Int64: Number of cells in domain, size of solution memory per variable.

r::Vector{Float64}: Radial coordinates of cells centers.

w::Vector{Float64}: Radial coordinates of cells walls.

DryTooling.FiniteVolumes.geometriccellsgrid1D — Function

geometriccellsgrid(R::Float64, N::Int64)::UserDefinedGrid1D

Helper for creating a UserDefinedGrid1D with all nodes equidistant.

DryTooling.FiniteVolumes.equidistantcellsgrid1D — Function

equidistantcellsgrid(R::Float64, N::Int64)::UserDefinedGrid1D

Helper for creating a UserDefinedGrid1D with all nodes equidistant.

Heat conduction

Mass transfer

DryTooling.FiniteVolumes.plotcarburizedprofile — Function

plotcarburizedprofile(
	model::AusteniteCarburizing1DModel,
	yc0::Float64;
	showstairs = true,
	xticks = nothing,
	yticks = nothing,
    label = nothing
)

Display carburized profile and mass intake in a standardized way.

Theory guide

Please go to the module theory guide page.

Models validation

Iron carburizing