DryTooling.Granular

General porous media

DryTooling.Granular.PackedBedPorosityDescriptor — Type

Provides description of porosity parameters with stochastic behavior.

ϕ::Union{Float64, Vector{Float64}}: Porosity volume fraction in medium [-].

l::Union{Float64, Vector{Float64}}: Characteristic particle size in medium [m].

σϕ::Union{Nothing, Float64}: Optional standard deviation of porosity volume fraction [-].

σl::Union{Nothing, Float64}: Optional standard deviation of characteristic particle size [m].

P::Union{Float64, Vector{Float64}}: Perimeter in reactor cross-section [m].

D::Union{Float64, Vector{Float64}}: Characteristic diameter of porosity channels [m].

A::Float64: Reactor area used for scaling perimeter [m²].

Usage

PackedBedPorosityDescriptor can be used to describe the geometry of exchange section of a packed bed for a single set of arguments.

julia> using DryTooling.Granular

julia> PackedBedPorosityDescriptor(; ϕ = 0.65, l = 0.10, area = 1.0)
PackedBedPorosityDescriptor(P = 21.000000 m, D = 0.123810 m)

It can also be used to describe randomly varying reactors, what is a more realistic thing to do when using this structure to simulate real world systems.

julia> using DryTooling.Granular

julia> PackedBedPorosityDescriptor(;
            ϕ  = 0.65, l  = 0.10,
            σϕ = 0.03, σl = 0.01,
            N = 2,
            ϕlims = (0.4, 0.8),
            llims = (0.0, 0.3),
            seed = 42,
            area = 1.0
        )
PackedBedPorosityDescriptor(
    P from  21.455749 m to  24.370742 m
    D from   0.125589 m to   0.102353 m
)

Rotary kiln models

The structure SymbolicLinearKramersModel implements the Kramers' ordinary differential equation for prediction of bed height profile in a rotary kiln. This equation is implemented under the formalism of ModelingToolkit.

DryTooling.Granular.SymbolicLinearKramersModel — Type

Creates a reusable linear Kramers model for rotary kiln simulation.

R::Symbolics.Num: Symbolic kiln internal radius

Φ::Symbolics.Num: Symbolic kiln feed rate

ω::Symbolics.Num: Symbolic kiln rotation rate

β::Symbolics.Num: Symbolic kiln slope

γ::Symbolics.Num: Symbolic solids repose angle

z::Symbolics.Num: Symbolic kiln axial coordinates

h::Symbolics.Num: Symbolic bed height profile

sys::ModelingToolkit.ODESystem: Problem ordinary differential equation

Description of a rotary kiln bed geometry computed from the solution of bed height along the kiln length. The main goal of the quantities computed here is their use with heat and mass transfer models for the simulation of rotary kiln process.

DryTooling.Granular.RotaryKilnBedSolution — Type

General geometric description of a bed from Kramers equation solution.

z::Vector{Float64}: Solution coordinates [m]

h::Vector{Float64}: Solution bed height [m]

θ::Vector{Float64}: View angle from kiln center [rad]

l::Vector{Float64}: Bed-freeboard cord length [m]

A::Vector{Float64}: Local bed cross section area [m²]

η::Vector{Float64}: Local loading based on height [-]

ηₘ::Float64: Mean loading of kiln [%]

V::Float64: Bed integral volume [m³]

τ::Float64: Residence time of particles

β::Float64: Kiln slope [rad]

Arguments

Internal elements are initialized through the following constructor:

RotaryKilnBedSolution(z::Vector{Float64}, h::Vector{Float64}, R::Float64, Φ::Float64)

Where parameters are given as:

z: solution coordinates over length, [m].

h: bed profile solution over length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

An outer constructor is also provided for managing the integration of an instance of SymbolicLinearKramersModel. This is the recommended usage that is illustrated below.

Important: inputs must be provided in international system (SI) units as a better physical practice. The only exception is the rotation rate ω provided in revolution multiples. If the discharge end is held by a dam, its height must be provided instead of the particle size, as it is used as the ODE initial condition.

model: a symbolic kiln model.

L: kiln length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

ω: kiln rotation rate, [rev/s].

β: kiln slope, [rad].

γ: solids repose angle, [rad].

d: particle size or dam height, [m].

solver: Solver for DifferentialEquations. Defaults to Tsit5.

rtol: Relative integration tolerance. Defaults to 1.0e-08.

atol: Absolute integration tolerance. Defaults to 1.0e-08.

Usage

Data in next example is an SI conversion of an example from Kramers (1952).

julia> using DryTooling.Granular

julia> L = 13.715999999999998;  # Kiln length [m]

julia> D = 1.8897599999999999;  # Kiln diameter [m]

julia> β = 2.3859440303888126;  # Kiln slope [°]

julia> γ = 45.0;                # Repose angle [°]

julia> d = 1.0;                 # Particle/dam size [mm]

julia> Φ = 10.363965852671996;  # Feed rate [m³/h]

julia> ω = 3.0300000000000002;  # Rotation rate [rev/min]

julia> bed = RotaryKilnBedSolution(;
            model = SymbolicLinearKramersModel(),
            L     = L,
            R     = D / 2.0,
            Φ     = Φ / 3600.0,
            ω     = ω / 60.0,
            β     = deg2rad(β),
            γ     = deg2rad(γ),
            d     = d / 1000.0
        );

julia> bed
RotaryKilnBedSolution(τ = 13.169938 min, ηₘ = 5.913271 %)

julia> bed.τ
790.1963002204092

DryTooling.Granular.plotlinearkramersmodel — Function

plotlinearkramersmodel(
    model::RotaryKilnBedSolution;
    normz::Bool = false,
    normh::Bool = false
)::Figure

Standardized plotting of RotaryKilnBedSolution bed profile. It supports normalization of axes throught keywords normz for axial coordinate and normh for bed depth.

Finally a set of basic equations provided for process analysis.

DryTooling.Granular.sullivansηₘ — Function

Sullivans approximation to kiln filling.

DryTooling.Granular.dimlessNΦ — Function

Kramers (1952) dimensionless group NΦ.

DryTooling.Granular.dimlessNₖ — Function

Kramers (1952) dimensionless group Nₖ.

DryTooling.Granular.perrayresidence — Function

Compute residence time from Peray's equation.

DryTooling.Granular.kramersnlapprox — Function

Nonlinear formulation of Kramers model approximate solution.

Theory guide

Please go to the module theory guide page.

Models validation

Kramers' model