DryTooling.Simulation

NOTE: this module is fragile and breaking changes are still expected. It is not until all the main solvers are migrated that it will become stable. This is necessary for ensuring compatibility with all models.

Iterative solver

The core of the iterative time-stepping solver is step!. This function that is described below works according to the following solution logic:

The fouter! update function is called once. This is generall where one implements the right-hand side update of the problem before stepping.

An initial update with finner! is done. Normally this is responsible by the update of matrix coefficients that are dependent on solution state.

If relaxation α <= 0.0, then the problem is treated as linear.

Otherwise a maximum of M iterations are repeated, where fsolve! is used to solve the problem (ofter an under-relaxation step) and is also expected to keep track of residuals.

Problem coefficients are updated with finner! if not converged.

DryTooling.Simulation.step! — Function

step!(
    m::AbstractPhysicalModel,
    t::Float64,
    n::Int64;
    fouter!::Function,
    finner!::Function,
    fsolve!::Function,
    α::Float64 = 0.1,
    iters::Int64 = 20,
    tol::Float64 = 1.0e-10
)

Manage the integration of a model m from time t corresponding to step call n using model internal time step. All the updates of coefficients and solution are performed through user-supplied functions.

The outer iteration for advancing between steps is carried out by advance!.

DryTooling.Simulation.advance! — Function

advance!(
    m::AbstractPhysicalModel;
    α::Float64 = 0.1,
    ε::Float64 = 1.0e-10,
    M::Int64 = 20
)

Manage execution of step! over the integration time interval.

Any model willing to implement its solution through the methods provided in this module is expected to explicity import and override the behaviour of the following methods for its own type:

DryTooling.Simulation.fouter! — Function

fouter!(::AbstractPhysicalModel, ::Float64, ::Int64)

Outer loop update for step!.

DryTooling.Simulation.finner! — Function

finner!(::AbstractPhysicalModel, ::Float64, ::Int64)

Inner loop update for step!.

DryTooling.Simulation.fsolve! — Function

fsolve!(::AbstractPhysicalModel, ::Float64, ::Int64, ::Float64)

Solution update for step!.

DryTooling.Simulation.timepoints — Function

timepoints(::AbstractPhysicalModel)

Get array of model time-points for use in step!.

Linear algebra

DryTooling.Simulation.TridiagonalProblem — Type

Memory for a tridiagonal problem of rank N.

All tensors are filled with zeros upon creation. This is simply a utility for memory allocation, no other operations are made.

A::LinearAlgebra.Tridiagonal{Float64, Vector{Float64}}: Main problem matrix.

b::Vector{Float64}: Right-hand side vector.

x::Vector{Float64}: Solution variable vector.

a::Vector{Float64}: Auxiliary vector, e.g. for iterative problems.

DryTooling.Simulation.solve! — Method

solve!(p::TridiagonalProblem)::Nothing

Solve a tridiagonal problem and update internal memory.

DryTooling.Simulation.change — Function

change(p::TridiagonalProblem)::Vector{Float64}

Compute change in solution of a tridiagonal problem without update.

Residuals tracking

DryTooling.Simulation.TimeSteppingSimulationResiduals — Type

Manage time-stepping solvers residuals storage during a simulation.

The memory is initialized with a given number of inner and outer iterations and resizing is not under the scope of this structure.

N::Int64: Number of variables being tracked in problem.

counter::Base.RefValue{Int64}: Total iteration counter.

innersteps::Vector{Int64}: Number of inner steps per outer loop in solution.

residuals::Matrix{Float64}: Store residuals of each inner step, one variable per column.

Usage

For starting a simulation, use the outer constructor for starting a simulation with pre-allocated memory with interface:

TimeSteppingSimulationResiduals(N::Int64, inner::Int64, outer::Int64)

Once the simulation is finished, the first instance can be processed through creation of a new object using the next interface:

TimeSteppingSimulationResiduals(r::TimeSteppingSimulationResiduals)

DryTooling.Simulation.finaliterationdata — Function

finaliterationdata(
    r::TimeSteppingSimulationResiduals
)::Tuple{Vector{Int64}, Matrix{Float64}}

Retrieve data at iterations closing an outer loop of solution.

DryTooling.Simulation.addresidual! — Function

addresidual!(
    r::TimeSteppingSimulationResiduals,
    ε::Vector{Float64}
)::Nothing

Utility to increment iteration counter and store residuals.

DryTooling.Simulation.plotsimulationresiduals — Function

plotsimulationresiduals(
    r::TimeSteppingSimulationResiduals;
    ε::Union{Float64, Nothing} = nothing,
    showinner::Bool = false,
    resolution::Tuple{Int64, Int64} = (720, 500)
)::Tuple{Figure, Axis, Vector}

Plot problem residuals over iterations or steps. It performs the basic figure setup, configuration of axis and details beign left to the user.

Examples

Please go to the module samples page.