DryTooling API

DryTooling.GAS_CONSTANT — Constant

Ideal gas constant [8.31446261815324 $J mol^{-1} K^{-1}$].

DryTooling.ONE_ATM — Constant

Atmospheric pressure at sea level [101325.0 $Pa$].

DryTooling.STABLE_ELEMENTS_TABLE — Constant

Instantiation of stable elements table.

DryTooling.STEFAN_BOLTZMANN — Constant

Stefan-Boltzmann constant [5.670374419e-8 $W m^{-2} K^{-4}$]

DryTooling.TRANSPORT_MODELS — Constant

Named access to transport models.

DryTooling.ZERO_CELSIUS — Constant

Zero degrees Celsius in Kelvin [273.15 $K$].

DryTooling.AbstractDiffusionModel1D — Type

Base type for diffusion (heat, species, ...) models.

DryTooling.AbstractGasThermo — Type

Base type for thermodynamic models.

DryTooling.AbstractGrid1D — Type

Base type of one-dimensional grids.

DryTooling.AbstractIterativeSolver — Type

Base type for (nonlinear) iterative solvers.

DryTooling.AbstractKineticsMechanism — Type

Base type for coded kinetics mechanisms.

DryTooling.AbstractMatrixProblem — Type

Base type for linear algebra problems.

DryTooling.AbstractMixturePhase — Type

Base type for simplified mixture phases.

DryTooling.AbstractMixtureSubstance — Type

Base type for simplified mixture substances.

DryTooling.AbstractPhysicalModel — Type

Base type for physical models.

DryTooling.AbstractSolidMaterial — Type

Base type for solid materials.

DryTooling.AbstractSolidThermo — Type

Base type for thermodynamic models.

DryTooling.AbstractSolidTransport — Type

Base type for transport models.

DryTooling.AbstractSolutionStorage — Type

Base type for storing simulation solution.

DryTooling.AbstractTransportModel — Type

Base type for transport models.

DryTooling.axesunitscaler — Method

axesunitscaler(x::Number)::Tuple{String, Int64}

Find scaling factor for multiples of 1000 units. Together with closestpowerofx this can be used to produce better automatic plot axes limits. The returned values provide the string for modifying the axis label and the associated scaling factor.

NOTE: this function is not yet stable. In the future it will instead return labels using symbols like k, M, G for the units through a flag provided by the user.

julia> axesunitscaler(1)
("", 1)
julia> axesunitscaler(1000)
("[×1000]", 1000)
julia> axesunitscaler(1000000)
("[×1000000]", 1000000)

DryTooling.body — Method

body(z)

Access view of array body. See also head and tail.

julia> body([1, 2, 3, 4])
2-element view(::Vector{Int64}, 2:3) with eltype Int64:
 2
 3
julia> body(1:4)
2:3

DryTooling.closestpowerofx — Method

closestpowerofx(
    v::Number;
    x::Number = 10,
    roundf::Function = ceil
)::Int64

Compute the integer power of x closest to v using roundf as rouding method. This might be useful for automatic setting more reasonable limits to plot axis or similar applications. Changing the rouding method through roundf is also possible.

julia> closestpowerofx(12.0; x = 10)
20
julia> closestpowerofx(12.0; x = 10, roundf = floor)
10
julia> closestpowerofx(12.0; x = 10, roundf = round)
10

DryTooling.head — Method

head(z)

Access view of array head. See also $tail$ and $body$.

julia> head(1:4)
1:3
julia> head([1, 2, 3, 4])
3-element view(::Vector{Int64}, 1:3) with eltype Int64:
 1
 2
 3

DryTooling.heaviside — Method

heaviside(t)

Provides a Heaviside function compatible with automatic differentiation. This is a requirement for conceiving, e.g., model predictive controls with discontinuous functions under ModelingToolkit.

Usage

julia> heaviside(-2:2)
5-element Vector{Float64}:
 0.0
 0.0
 0.5
 1.0
 1.0

DryTooling.interval — Method

interval(x; a=-Inf, b=Inf)

Returns 1 if $x ∈ (a, b)$, 1/2 for $x = a || x = b$, or 0 .

Usage

julia> interval(0:6; a = 2, b = 5)
7-element Vector{Float64}:
 0.0
 0.0
 0.5
 1.0
 1.0
 0.5
 0.0

DryTooling.makestepwise1d — Method

makestepwise1d(lo, hi, xc)

Creates an univariate function that is composed of two parts, the first evaluated before a critical domain point xc, and the second above that value. This is often required, for instance, for the evaluation of NASA polynomials for thermodynamic properties. If differentiable, then the returned function is compatible with symbolic argument as required when using package ModelingToolkit, etc.

Usage

julia> f = makestepwise1d(x->x, x->x^2, 1.0; differentiable = true);

julia> f(0:0.2:2.0)
11-element Vector{Float64}:
 0.0
 0.2
 0.4
 0.6
 0.8
 1.0
 1.44
 1.9599999999999997
 2.5600000000000005
 3.24
 4.0

julia> using ModelingToolkit

julia> @variables x
1-element Vector{Num}:
 x

julia> h(x); # Output is too long, try by yourself.

DryTooling.tail — Method

tail(z)

Access view of array tail. See also head and body.

julia> tail([1, 2, 3, 4])
3-element view(::Vector{Int64}, 2:4) with eltype Int64:
 2
 3
 4
julia> tail(1:4)
2:4

Index

DryTooling.GAS_CONSTANT

DryTooling.ONE_ATM

DryTooling.STABLE_ELEMENTS_TABLE

DryTooling.STEFAN_BOLTZMANN

DryTooling.TRANSPORT_MODELS

DryTooling.ZERO_CELSIUS

DryTooling.AbstractDiffusionModel1D

DryTooling.AbstractGasThermo

DryTooling.AbstractGrid1D

DryTooling.AbstractIterativeSolver

DryTooling.AbstractKineticsMechanism

DryTooling.AbstractMatrixProblem

DryTooling.AbstractMixturePhase

DryTooling.AbstractMixtureSubstance

DryTooling.AbstractPhysicalModel

DryTooling.AbstractSolidMaterial

DryTooling.AbstractSolidThermo

DryTooling.AbstractSolidTransport

DryTooling.AbstractSolutionStorage

DryTooling.AbstractTransportModel

DryTooling.axesunitscaler

DryTooling.body

DryTooling.closestpowerofx

DryTooling.head

DryTooling.heaviside

DryTooling.interval

DryTooling.makestepwise1d

DryTooling.tail