Granular

DryTooling.Granular.PackedBedPorosityDescriptor — Type

Provides description of porosity parameters with stochastic behavior.

ϕ::Union{Float64, Vector{Float64}}: Porosity volume fraction in medium [-].

l::Union{Float64, Vector{Float64}}: Characteristic particle size in medium [m].

σϕ::Union{Nothing, Float64}: Optional standard deviation of porosity volume fraction [-].

σl::Union{Nothing, Float64}: Optional standard deviation of characteristic particle size [m].

P::Union{Float64, Vector{Float64}}: Perimeter in reactor cross-section [m].

D::Union{Float64, Vector{Float64}}: Characteristic diameter of porosity channels [m].

A::Float64: Reactor area used for scaling perimeter [m²].

Some background

Modeling of geometrical characteristics of porous beds is required for including both their thermal effect or role over chemistry in chemical reactors. A classical approach used in several commercial and open source tools is that of Gunn (1978) [Gunn1978]. In what follows we develop the ideas that lead to an analogous model which is implemented by this structure.

To build the model we will assume a reactor of constant rectangular cross-section ${A}_{r}={b}{w}$ and volume ${V}_{R}={b}{w}{h}$. Its cross-section perimeter is then ${P}_{R}=2({b}+{w})$. Inside this reactor we randomly pack cubic particles $\beta$ of surface area ${A}_{\beta}=6{l}_{\beta}^2$ and volume ${V}_{\beta}={l}_{\beta}^3$ at a porosity level ${\phi}$. Thus the total volume of solids inside the reactor is ${V}_{S}=(1-{\phi}){V}_{R}$ and the approximate number of particles ${N}=\frac{{V}_{S}}{{V}_{\beta}}$. Following a similar reasoning the total surface area of particles is ${A}_{S}={N}{A}_{\beta}$. Performing all the substitutions so far one finds the following expression

\[{A}_{S}=\frac{6(1-{\phi}){b}{w}{h}}{{l}_{\beta}}\]

Since the differential $d{A}={P}d{l}$ holds for the surface of a body over its length ${l}$, one can divide the above expression by the reactor length to get the perimeter of particles in a cross-section. We can further divide by the cross-section area itself and find the perimeter density which is a more general result, and find the expression proposed by Gunn [Gunn1978]. This result is summarized in the next equation where the subscript of particle size was dropped for generality.

\[{P} = \frac{6(1-{\phi})}{{l}}\]

An estimator of the number of channels per unit cross-section of reactor ${N}$ can be related to the porosity through ${N}\pi{R}^2={phi}$. Because the above perimeter is shared between the fluid volume and solids, it holds that ${N}2\pi{R}=P$. Using these expressions one can solve for the porosity channels characteristic radius ${R}$ as given below, which is also a result reported by Gunn [Gunn1978].

\[{R}=\frac{{\phi}{l}}{3(1-{\phi})}\]

DryTooling.Granular.RotaryKilnBedSolution — Type

RotaryKilnBedSolution

Description of a rotary kiln bed geometry computed from the solution of bed height along the kiln length. The main goal of the quantities computed here is their use with heat and mass transfer models for the simulation of rotary kiln process.

Internal elements are initialized through the following constructor:

RotaryKilnBedSolution(
    z::Vector{Float64},
    h::Vector{Float64},
    R::Float64,
    Φ::Float64
)

Where parameters are given as:

- `z`: solution coordinates over length, [m].
- `h`: bed profile solution over length, [m].
- `R`: kiln internal radius, [m].
- `Φ`: kiln feed rate, [m³/s].

z::Vector{Float64}: Solution coordinates [m]

h::Vector{Float64}: Solution bed height [m]

θ::Vector{Float64}: View angle from kiln center [rad]

l::Vector{Float64}: Bed-freeboard cord length [m]

A::Vector{Float64}: Local bed cross section area [m²]

η::Vector{Float64}: Local loading based on height [-]

ηₘ::Float64: Mean loading of kiln [%]

V::Float64: Bed integral volume [m³]

τ::Float64: Residence time of particles

β::Float64: Kiln slope [rad]

DryTooling.Granular.SymbolicLinearKramersModel — Type

SymbolicLinearKramersModel

Creates a reusable linear Kramers model for rotary kiln simulation.

Implements the ordinary differential equation for prediction of bed height profile in a rotary kiln as proposed by Kramers and Croockewite (1952) [Kramers1952]. Its goal is to be used as a process support tool or to integrate more complex models requiring integration of the bed profile.

R::Symbolics.Num: Symbolic kiln internal radius

Φ::Symbolics.Num: Symbolic kiln feed rate

ω::Symbolics.Num: Symbolic kiln rotation rate

β::Symbolics.Num: Symbolic kiln slope

γ::Symbolics.Num: Symbolic solids repose angle

z::Symbolics.Num: Symbolic kiln axial coordinates

h::Symbolics.Num: Symbolic bed height profile

sys::ModelingToolkit.ODESystem: Problem ordinary differential equation

Index

DryTooling.Granular.PackedBedPorosityDescriptor

DryTooling.Granular.RotaryKilnBedSolution

DryTooling.Granular.SymbolicLinearKramersModel

Gunn1978D. J. Gunn, 1978

Kramers1952Kramers et al., 1952