DryTooling.Simulation

NOTE: this module is fragile and breaking changes are still expected. It is not until all the main solvers are migrated that it will become stable. This is necessary for ensuring compatibility with all models.

Residuals tracking

DryTooling.Simulation.TimeSteppingSimulationResiduals — Type

Manage time-stepping solvers residuals storage during a simulation.

The memory is initialized with a given number of inner and outer iterations and resizing is not under the scope of this structure.

N::Int64: Number of variables being tracked in problem.

counter::Base.RefValue{Int64}: Total iteration counter.

innersteps::Vector{Int64}: Number of inner steps per outer loop in solution.

residuals::Matrix{Float64}: Store residuals of each inner step, one variable per column.

Usage

For starting a simulation, use the outer constructor for starting a simulation with pre-allocated memory with interface:

TimeSteppingSimulationResiduals(N::Int64, inner::Int64, outer::Int64)

Once the simulation is finished, the first instance can be processed through creation of a new object using the next interface:

TimeSteppingSimulationResiduals(r::TimeSteppingSimulationResiduals)

DryTooling.Simulation.finaliterationdata — Function

finaliterationdata(
    r::TimeSteppingSimulationResiduals
)::Tuple{Vector{Int64}, Matrix{Float64}}

Retrieve data at iterations closing an outer loop of solution.

DryTooling.Simulation.addresidual! — Function

addresidual!(
    r::TimeSteppingSimulationResiduals,
    ε::Vector{Float64}
)::Nothing

Utility to increment iteration counter and store residuals.

DryTooling.Simulation.plotsimulationresiduals — Function

plotsimulationresiduals(
    r::TimeSteppingSimulationResiduals;
    ε::Union{Float64, Nothing} = nothing,
    showinner::Bool = false,
    resolution::Tuple{Int64, Int64} = (720, 500)
)::Tuple{Figure, Axis, Vector}

Plot problem residuals over iterations or steps. It performs the basic figure setup, configuration of axis and details beign left to the user.

Tutorial: residuals tracking in a solver

The residuals tracking functionalities of module DryTooling.Simulation` are not often imported by the end-user (except for its plotting utility function). In this tutorial we illustrate the logic of using a residual tracker in a new solver.

using DryTooling.Simulation

N = 2      # Number of variables.
M = 5      # Maximum inner steps.
steps = 10 # Time-advancement steps.

# Create a TimeSteppingSimulationResiduals object with the number of variables
# to track, how many inner iterations per step are expected, and the
# number of steps.
#
# IMPORTANT: If the total number of iterations is exceeded, it is up
# to the user to allocate more memory, the tracker will not manage it!
r = TimeSteppingSimulationResiduals(N, M, steps)

# The following loop represents a *dummy solver*. The outer loop
# provides the time-advancement while the inner loop handles the
# nonlinear problem. In the inner loop we use a random number
# generator to provide varying number of steps per outer step.
for kouter in 1:steps
    for kinner in 1:rand(2:M)
        # Keep track of inner iterations per step.
        r.innersteps[kouter] = kinner

        # Feed residuals to the solver.
        addresidual!(r, rand(r.N))
    end
end

# After running a simulation we create a new object using another
# constructor that accepts a `TimeSteppingSimulationResiduals` object. This
# handles the post-processing.
s = TimeSteppingSimulationResiduals(r)

# The new object is ready for visualization. Check the documentation
# of the following function for more details. It provides a raw figure
# and handles for modifying it for proper display (uncomment below).
# fig = plotsimulationresiduals(s; showinner = true)[1]