DryTooling

Documentation for DryTooling.

DryTooling.Kramers

DryTooling.Kramers.RotaryKilnBedSolution

DryTooling.Kramers.RotaryKilnBedSolution

DryTooling.Kramers.SymbolicLinearKramersModel

DryTooling.Kramers.SymbolicLinearKramersModel

DryTooling.Kramers.plotlinearkramersmodel

DryTooling.Kramers.solvelinearkramersmodel

DryTooling.Kramers.solvelinearkramersmodelSI

DryTooling.Kramers — Module

module Kramers

Implements the differential equation for prediction of bed height profile in a rotary kiln as proposed by Kramers and Croockewite (1952) [1].

DryTooling.Kramers.RotaryKilnBedSolution — Type

`RotaryKilnBedSolution`

Description of a rotary kiln bed geometry computed from the solution of bed height along the kiln length. The main goal of the quantities computed here is their use with heat and mass transfer models for the simulation of rotary kiln process.

DryTooling.Kramers.RotaryKilnBedSolution — Method

`RotaryKilnBedSolution(
    z::Vector{Float64},
    h::Vector{Float64},
    R::Float64,
    Φ::Float64
)`

z: solution coordinates over length, [m].

h: bed profile solution over length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

DryTooling.Kramers.SymbolicLinearKramersModel — Type

`SymbolicLinearKramersModel``

Creates a reusable linear Kramers model for rotary kiln simulation.

DryTooling.Kramers.SymbolicLinearKramersModel — Method

`SymbolicLinearKramersModel()`

Symbolic model constructor.

DryTooling.Kramers.plotlinearkramersmodel — Method

plotlinearkramersmodel(
    model::SolutionLinearKramersModel;
    normz::Bool = false,
    normh::Bool = false
)::Any

Display plot of model solution for rotary kiln bed profile. Arguments normz and normh control whether z-coordinate and bed height must be normalized, respectively.

DryTooling.Kramers.solvelinearkramersmodel — Method

`solvelinearkramersmodel(;
    L::Float64,
    D::Float64,
    Φ::Float64,
    ω::Float64,
    β::Float64,
    γ::Float64,
    d::Float64,
    model::Union{SymbolicLinearKramersModel,Nothing} = nothing
)::RotaryKilnBedSolution`

L: kiln length, [m].

D: kiln internal diameter, [m].

Φ: kiln feed rate, [m³/h].

ω: kiln rotation rate, [rev/min].

β: kiln slope, [°].

γ: solids repose angle, [°].

d: particle size or dam height, [mm].

Important: inputs are in customary units based in international system (SI) - no imperial units supported - while all the computed values are provided in SI units as a better physical practice. If d represents a particle size, take care that it is provided in millimeters for ease of exchange with common dam sizes.

DryTooling.Kramers.solvelinearkramersmodelSI — Method

`solvelinearkramersmodelSI(;
    model::SymbolicLinearKramersModel,
    L::Float64,
    R::Float64,
    Φ::Float64,
    ω::Float64,
    β::Float64,
    γ::Float64,
    d::Float64,
    solver::Any = Tsit5(),
    rtol::Float64 = 1.0e-08,
    atol::Float64 = 1.0e-08
)`

Integrates an instance of SymbolicLinearKramersModel.

Note: if the discharge end is hold by a dam, its height must be provided instead of the particle size, as it is used as the ODE initial condition.

model: a symbolic kiln model.

L: kiln length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

ω: kiln rotation rate, [rev/s].

β: kiln slope, [rad].

γ: solids repose angle, [rad].

d: particle size or dam height, [m].

1Kramers et al., 1952