DryTooling

Documentation for DryTooling.

I am often faced with using the same approach for different engineering and scientific problems, but I don't like repeating the same task again and again. This is where DryTooling.jl comes in. By adopting some principles of DRY in Julia, to a larger extent than its definition, it packages together models and workflows that are not available or validated elsewhere - and in some cases adapts existing models. The tools will progressively cover a broad range of numerical applications and data treatment, this package is in its early days from the migration of my old Python scripts and packages.

Also dry tooling is my favorite sport!

Citing

See CITATION.bib for the relevant reference(s).

Table of contents

DryTooling.GAS_CONSTANT

DryTooling.ONE_ATM

DryTooling.STABLE_ELEMENTS_TABLE

DryTooling.TRANSPORT_MODELS

DryTooling.ZERO_CELSIUS

DryTooling.AbstractGasThermo

DryTooling.AbstractSolidMaterial

DryTooling.AbstractSolidThermo

DryTooling.AbstractSolidTransport

DryTooling.AbstractTransportModel

DryTooling.ElementData

DryTooling.IdealGasMixture

DryTooling.IdealGasSpecies

DryTooling.IdealGasThermo

DryTooling.LennardJonesTransport

DryTooling.MaterialPowderBed

DryTooling.MaterialShomate

DryTooling.MaterialTransportProperties

DryTooling.RotaryKilnBedSolution

DryTooling.StableElementsTable

DryTooling.SymbolicLinearKramersModel

DryTooling.concentration

DryTooling.densitymass

DryTooling.element

DryTooling.elementmass

DryTooling.enthalpymass

DryTooling.enthalpymole

DryTooling.getnameditem

DryTooling.getthermo

DryTooling.heaviside

DryTooling.interval

DryTooling.makestepwise1d

DryTooling.mass

DryTooling.mass

DryTooling.massfraction2molefraction

DryTooling.meanmolecularmass

DryTooling.meanmolecularmass

DryTooling.molefraction2massfraction

DryTooling.nasa7enthapy

DryTooling.nasa7specificheat

DryTooling.plotlinearkramersmodel

DryTooling.shomateenthalpy

DryTooling.shomateentropy

DryTooling.shomatespecificheat

DryTooling.solvelinearkramersmodel

DryTooling.specificheatmass

DryTooling.specificheatmass

DryTooling.specificheatmole

DryTooling.GAS_CONSTANT — Constant

Ideal gas constant [J/(mol.K)].

DryTooling.ONE_ATM — Constant

Atmospheric pressure at sea level [Pa].

DryTooling.STABLE_ELEMENTS_TABLE — Constant

Instantiation of stable elements table.

DryTooling.TRANSPORT_MODELS — Constant

Named access to transport models.

DryTooling.ZERO_CELSIUS — Constant

Zero degrees Celsius in Kelvin for conversion [K].

DryTooling.AbstractGasThermo — Type

Base type for thermodynamic models.

DryTooling.AbstractSolidMaterial — Type

Base type for

DryTooling.AbstractSolidThermo — Type

Base type for thermodynamic models.

DryTooling.AbstractSolidTransport — Type

Base type for transport models.

DryTooling.AbstractTransportModel — Type

Base type for transport models.

DryTooling.ElementData — Type

Represents a chemical element.

DryTooling.IdealGasMixture — Type

Ideal gas phase mixture model.

DryTooling.IdealGasSpecies — Type

Ideal gas phase species model.

DryTooling.IdealGasThermo — Type

Ideal gas phase thermodynamics model.

DryTooling.LennardJonesTransport — Type

Lennard-Jones ideal gas transport model.

DryTooling.MaterialPowderBed — Type

MaterialPowderBed

Description of a powder bed material for a rotary kiln.

ρ::Float64: Density [kg/m³].

γ::Float64: Repose angle [rad].

ϕ::Float64: Solid packing fraction [-].

d::Float64: Particle mean diameter [m].

M::Float64: Molar mass [kg/mol].

k::Function: Thermal conductivity [W/(m.K)].

ε::Function: Emissivity [-].

cₚ::Function: Molar specific heat [J/(mol.K)].

h::Function: Molar enthalpy [J/mol].

s::Function: Molar entropy [J/K].

thermo::DryTooling.AbstractSolidThermo: Access to thermodynamic model.

transport::DryTooling.AbstractSolidTransport: Access to transport model.

DryTooling.MaterialShomate — Type

MaterialShomate

IMPORTANT: the implementation of makestepwise1d used for step-wise function evaluation takes the mean of both ranges at the T_ch, while the actual Shomate uses the high range for doing so. The function enforces the right behavior by multiplying T_ch the value by 1-eps().

cₚ::Function: Molar specific heat [J/(mol.K)].

h::Function: Molar enthalpy [J/mol].

s::Function: Molar entropy [J/K].

a_lo::Vector{Float64}: Low temperature range Shomate coefficients.

a_hi::Vector{Float64}: High temperature range Shomate coefficients.

T_ch::Float64: Temperature of range change for evaluation.

DryTooling.MaterialTransportProperties — Type

MaterialTransportPolynomial

Transport properties for a solid material.

k::Function: Thermal conductivity [W/(m.K)].

ε::Function: Emissivity [-].

DryTooling.RotaryKilnBedSolution — Type

RotaryKilnBedSolution

Description of a rotary kiln bed geometry computed from the solution of bed height along the kiln length. The main goal of the quantities computed here is their use with heat and mass transfer models for the simulation of rotary kiln process.

Internal elements are initialized through the following constructor:

RotaryKilnBedSolution(
    z::Vector{Float64},
    h::Vector{Float64},
    R::Float64,
    Φ::Float64
)

Where parameters are given as:

- `z`: solution coordinates over length, [m].
- `h`: bed profile solution over length, [m].
- `R`: kiln internal radius, [m].
- `Φ`: kiln feed rate, [m³/s].

z::Vector{Float64}: Solution coordinates [m]

h::Vector{Float64}: Solution bed height [m]

θ::Vector{Float64}: View angle from kiln center [rad]

l::Vector{Float64}: Bed-freeboard cord length [m]

A::Vector{Float64}: Local bed cross section area [m²]

η::Vector{Float64}: Local loading based on height [-]

ηₘ::Float64: Mean loading of kiln [%]

V::Float64: Bed integral volume [m³]

τ::Float64: Residence time of particles

β::Float64: Kiln slope [rad]

DryTooling.StableElementsTable — Type

Periodic table of stable chemical elements.

DryTooling.SymbolicLinearKramersModel — Type

SymbolicLinearKramersModel

Creates a reusable linear Kramers model for rotary kiln simulation.

Implements the ordinary differential equation for prediction of bed height profile in a rotary kiln as proposed by Kramers and Croockewite (1952) [1]. Its goal is to be used as a process support tool or to integrate more complex models requiring integration of the bed profile.

R::Symbolics.Num: Symbolic kiln internal radius

Φ::Symbolics.Num: Symbolic kiln feed rate

ω::Symbolics.Num: Symbolic kiln rotation rate

β::Symbolics.Num: Symbolic kiln slope

γ::Symbolics.Num: Symbolic solids repose angle

z::Symbolics.Num: Symbolic kiln axial coordinates

h::Symbolics.Num: Symbolic bed height profile

sys::ModelingToolkit.ODESystem: Problem ordinary differential equation

DryTooling.concentration — Method

Mixture concentration [mol/m³].

DryTooling.densitymass — Method

Mixture specific mass [kg/m³].

DryTooling.element — Method

Retrieve an element by name.

DryTooling.elementmass — Method

Retrieve atomic mass of element from atomic symbol [g/mol].

DryTooling.enthalpymass — Method

Species enthalpy in mass units [J/kg].

DryTooling.enthalpymole — Method

Species enthalpy in mole units [J/mol].

DryTooling.getnameditem — Method

Query first item matching name in dictionary.

DryTooling.getthermo — Method

Create specific heat and enthalpy functions for species.

DryTooling.heaviside — Method

Automatic differentiable Heaviside function.

DryTooling.interval — Method

Returns 1 if $x ∈ (a, b)$, 1/2 for $x = a || x = b$, or 0 .

DryTooling.makestepwise1d — Method

makestepwise1d(lo, hi, xc)

Creates an univariate function that is composed of two parts, the first evaluated before a critical domain point xc, and the seconda above that value. This is often required, for instance, for the evaluation of NASA polynomials for thermodynamic properties. If differentiable, then the returned function is compatible with symbolic argument as required when using package ModelingToolkit, etc.

DryTooling.mass — Method

Retrieve atomic mass of element [g/mol].

DryTooling.mass — Method

Retrieve atomic mass of species [kg/mol].

DryTooling.massfraction2molefraction — Method

Convert mass fractions to mole fractions.

DryTooling.meanmolecularmass — Method

Mixture mean molecular mass [kg/mol].

DryTooling.meanmolecularmass — Method

Mixture mean molecular mass [kg/mol].

DryTooling.molefraction2massfraction — Method

Convert mole fractions to mass fractions.

DryTooling.nasa7enthapy — Method

Molar enthalpy from NASA7 polynomial [J/mol].

DryTooling.nasa7specificheat — Method

Molar specific heat from NASA7 polynomial [J/(mol.K)].

DryTooling.plotlinearkramersmodel — Method

plotlinearkramersmodel(
    model::SolutionLinearKramersModel;
    normz::Bool = false,
    normh::Bool = false
)::Any

Display plot of model solution for rotary kiln bed profile. Arguments normz and normh control whether z-coordinate and bed height must be normalized, respectively.

DryTooling.shomateenthalpy — Method

shomateenthalpy(T::Float64, c::Vector{Float64})::Float64

Molar enthalpy with Shomate equation [J/mol].

DryTooling.shomateentropy — Method

shomateentropy(T::Float64, c::Vector{Float64})::Float64

Entropy with Shomate equation [J/K].

DryTooling.shomatespecificheat — Method

shomatespecificheat(T::Float64, c::Vector{Float64})::Float64

Molar specific heat with Shomate equation [J/(mol.K)].

DryTooling.solvelinearkramersmodel — Method

solvelinearkramersmodel(;
    model::SymbolicLinearKramersModel,
    L::Float64,
    R::Float64,
    Φ::Float64,
    ω::Float64,
    β::Float64,
    γ::Float64,
    d::Float64,
    solver::Any = Tsit5(),
    rtol::Float64 = 1.0e-08,
    atol::Float64 = 1.0e-08
)

Integrates an instance of SymbolicLinearKramersModel.

Important: inputs must be provided in international system (SI) units as a better physical practice. The only exception is the rotation rate ω provided in revolution multiples. If the discharge end is held by a dam, its height must be provided instead of the particle size, as it is used as the ODE initial condition.

model: a symbolic kiln model.

L: kiln length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

ω: kiln rotation rate, [rev/s].

β: kiln slope, [rad].

γ: solids repose angle, [rad].

d: particle size or dam height, [m].

DryTooling.specificheatmass — Method

Mixture mass-averaged specific heat [J/(kg.K)].

DryTooling.specificheatmass — Method

Species specific heat in mass units [J/(kg.K)].

DryTooling.specificheatmole — Method

Species specific heat in mole units [J/(mol.K)].

1Kramers et al., 1952