DryToolingCore

The core package of DryTooling ecosystem provides shared functionalities and types that are used in several other more specialized packages. This allows for standardization of interfaces, employed quantities, and avoid boilerplate code. Rather than documenting the exposed functionalities in implemented order, this page organizes everything to facilitate the understanding of the end-user. Here you find the constants by multi-purpose functionalities. All abstract types were grouped in a dedicated page.

Physical constants

DryToolingCore.GAS_CONSTANT — Constant

Ideal gas constant [8.31446261815324 $J mol^{-1} K^{-1}$].

DryToolingCore.ZERO_CELSIUS — Constant

Zero degrees Celsius in Kelvin [273.15 $K$].

DryToolingCore.ONE_ATM — Constant

Atmospheric pressure at sea level [101325.0 $Pa$].

DryToolingCore.STEFAN_BOLTZMANN — Constant

Stefan-Boltzmann constant [5.670374419e-8 $W m^{-2} K^{-4}$]

Haskell-like array slicing

Those who know Haskell probably started learning it by manipulating lists with head and tail. Those functionalities are not available in Julia by default and array slicing - with an ugly syntax - is required. Since this is done often in the fields of application of DryTooling, both head and tail together with a body functions are available in its core. They are simple wrapers over the @view macro and work with both iterable types and arrays. The following snippet illustrates their usage.

julia> v = collect(1:4);

julia> head(v) == [1; 2; 3]
true

julia> tail(v) == [2; 3; 4]
true

julia> body(v) == [2; 3]
true

More examples are provided in the following documentation ahead of each of the functions.

julia> head(1:4)
1:3

julia> head([1, 2, 3, 4])
3-element view(::Vector{Int64}, 1:3) with eltype Int64:
 1
 2
 3

DryToolingCore.head — Function

head(z)

Access view of array head. See also tail and body.

julia> tail([1, 2, 3, 4])
3-element view(::Vector{Int64}, 2:4) with eltype Int64:
 2
 3
 4
julia> tail(1:4)
2:4

DryToolingCore.tail — Function

tail(z)

Access view of array tail. See also head and body.

julia> body([1, 2, 3, 4])
2-element view(::Vector{Int64}, 2:3) with eltype Int64:
 2
 3
julia> body(1:4)
2:3

DryToolingCore.body — Function

body(z)

Access view of array body. See also head and tail.

Handling of discontinuous functions

Discontinuous functions are all over in real world applications. Whether they handle discrete signals sent to controllers or represent a material property change in the solution domain of a heat transfer simulation, they are often represented by a single or a composition of Heaviside step functions. Again, because its implementation is pretty simple and optimization routines require a differentiable form of this function, DryTooling implements heaviside and interval as proposed in this StackOverflow answer.

julia> heaviside(-1) == 0
true

julia> heaviside(-1.0) == 0.0
true

julia> heaviside(0.0) == 0.5
true

julia> heaviside(1.0) == 1.0
true

julia> interval(10; a = 0, b = 10) == 0.5
true

DryToolingCore.heaviside — Function

heaviside(t)

Provides a Heaviside function compatible with automatic differentiation. This is a requirement for conceiving, e.g., model predictive controls with discontinuous functions under ModelingToolkit.

Usage

julia> heaviside(-2:2)
5-element Vector{Float64}:
 0.0
 0.0
 0.5
 1.0
 1.0

DryToolingCore.interval — Function

interval(x; a=-Inf, b=Inf)

Returns 1 if $x ∈ (a, b)$, 1/2 for $x = a || x = b$, or 0 .

Usage

julia> interval(0:6; a = 2, b = 5)
7-element Vector{Float64}:
 0.0
 0.0
 0.5
 1.0
 1.0
 0.5
 0.0

DryToolingCore.makestepwise1d — Function

makestepwise1d(lo, hi, xc)

Creates an univariate function that is composed of two parts, the first evaluated before a critical domain point xc, and the second above that value. This is often required, for instance, for the evaluation of NASA polynomials for thermodynamic properties. If differentiable, then the returned function is compatible with symbolic argument as required when using package ModelingToolkit, etc.

Usage

julia> f = makestepwise1d(x->x, x->x^2, 1.0; differentiable = true);

julia> f(0:0.2:2.0)
11-element Vector{Float64}:
 0.0
 0.2
 0.4
 0.6
 0.8
 1.0
 1.44
 1.9599999999999997
 2.5600000000000005
 3.24
 4.0

julia> using ModelingToolkit

julia> @variables x
1-element Vector{Num}:
 x

julia> h(x); # Output is too long, try by yourself.

Rounding numbers and automatic axes

DryToolingCore.closestpowerofx — Function

closestpowerofx(
    v::Number;
    x::Number = 10,
    roundf::Function = ceil
)::Int64

Compute the integer power of x closest to v using roundf as rouding method. This might be useful for automatic setting more reasonable limits to plot axis or similar applications. Changing the rouding method through roundf is also possible.

julia> closestpowerofx(10)
10

julia> closestpowerofx(11)
20

julia> closestpowerofx(11, roundf = floor)
10

julia> closestpowerofx(11, x = 5, roundf = floor)
10

julia> closestpowerofx(12.0; x = 10)
20

julia> closestpowerofx(12.0; x = 10, roundf = floor)
10

julia> closestpowerofx(12.0; x = 10, roundf = round)
10

DryToolingCore.axesunitscaler — Function

axesunitscaler(x::Number)::Tuple{String, Int64}

Find scaling factor for multiples of 1000 units. Together with closestpowerofx this can be used to produce better automatic plot axes limits. The returned values provide the string for modifying the axis label and the associated scaling factor.

NOTE: this function is not yet stable. In the future it will instead return labels using symbols like k, M, G for the units through a flag provided by the user.

julia> axesunitscaler(1)
("", 1)
julia> axesunitscaler(1000)
("[×1000]", 1000)
julia> axesunitscaler(1000000)
("[×1000000]", 1000000)

Computation of changes and residuals

DryToolingCore.maxabsolutechange — Function

Maximum absolute change in a solution array.

DryToolingCore.maxrelativechange — Function

Maximum relative change in a solution array.

Index

DryToolingCore.GAS_CONSTANT

DryToolingCore.ONE_ATM

DryToolingCore.STEFAN_BOLTZMANN

DryToolingCore.ZERO_CELSIUS

DryToolingCore.AbstractDiffusionModel1D

DryToolingCore.AbstractGasThermo

DryToolingCore.AbstractGrid1D

DryToolingCore.AbstractIterativeSolver

DryToolingCore.AbstractKineticsMechanism

DryToolingCore.AbstractMatrixProblem

DryToolingCore.AbstractMixturePhase

DryToolingCore.AbstractMixtureSubstance

DryToolingCore.AbstractPhysicalModel

DryToolingCore.AbstractSolidMaterial

DryToolingCore.AbstractSolidThermo

DryToolingCore.AbstractSolidTransport

DryToolingCore.AbstractSolutionStorage

DryToolingCore.AbstractTransportModel

DryToolingCore.axesunitscaler

DryToolingCore.body

DryToolingCore.closestpowerofx

DryToolingCore.head

DryToolingCore.heaviside

DryToolingCore.interval

DryToolingCore.makestepwise1d

DryToolingCore.maxabsolutechange

DryToolingCore.maxrelativechange

DryToolingCore.tail