DryTooling.Kramers

Implements the ordinary differential equation for prediction of bed height profile in a rotary kiln as proposed by Kramers and Croockewite (1952) [1]. Its goal is to be used as a process support tool or to integrate more complex models requiring integration of the bed profile.

DryTooling.Kramers.RotaryKilnBedSolution

DryTooling.Kramers.RotaryKilnBedSolution

DryTooling.Kramers.SymbolicLinearKramersModel

DryTooling.Kramers.SymbolicLinearKramersModel

DryTooling.Kramers.plotlinearkramersmodel

DryTooling.Kramers.solvelinearkramersmodel

DryTooling.Kramers.RotaryKilnBedSolution — Type

RotaryKilnBedSolution

Description of a rotary kiln bed geometry computed from the solution of bed height along the kiln length. The main goal of the quantities computed here is their use with heat and mass transfer models for the simulation of rotary kiln process.

z::Vector{Float64}: Solution coordinates [m]

h::Vector{Float64}: Solution bed height [m]

θ::Vector{Float64}: View angle from kiln center [rad]

l::Vector{Float64}: Bed-freeboard cord length [m]

A::Vector{Float64}: Local bed cross section area [m²]

η::Vector{Float64}: Local loading based on height [-]

ηₘ::Float64: Mean loading of kiln [%]

V::Float64: Bed integral volume [m³]

τ::Float64: Residence time of particles

DryTooling.Kramers.RotaryKilnBedSolution — Method

RotaryKilnBedSolution(
    z::Vector{Float64},
    h::Vector{Float64},
    R::Float64,
    Φ::Float64
)

z: solution coordinates over length, [m].

h: bed profile solution over length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

DryTooling.Kramers.SymbolicLinearKramersModel — Type

SymbolicLinearKramersModel

Creates a reusable linear Kramers model for rotary kiln simulation.

R::Symbolics.Num: Symbolic kiln internal radius

Φ::Symbolics.Num: Symbolic kiln feed rate

ω::Symbolics.Num: Symbolic kiln rotation rate

β::Symbolics.Num: Symbolic kiln slope

γ::Symbolics.Num: Symbolic solids repose angle

z::Symbolics.Num: Symbolic kiln axial coordinates

h::Symbolics.Num: Symbolic bed height profile

sys::ModelingToolkit.ODESystem: Problem ordinary differential equation

DryTooling.Kramers.SymbolicLinearKramersModel — Method

SymbolicLinearKramersModel()

Symbolic model constructor.

DryTooling.Kramers.plotlinearkramersmodel — Method

plotlinearkramersmodel(
    model::SolutionLinearKramersModel;
    normz::Bool = false,
    normh::Bool = false
)::Any

Display plot of model solution for rotary kiln bed profile. Arguments normz and normh control whether z-coordinate and bed height must be normalized, respectively.

DryTooling.Kramers.solvelinearkramersmodel — Method

solvelinearkramersmodel(;
    model::SymbolicLinearKramersModel,
    L::Float64,
    R::Float64,
    Φ::Float64,
    ω::Float64,
    β::Float64,
    γ::Float64,
    d::Float64,
    solver::Any = Tsit5(),
    rtol::Float64 = 1.0e-08,
    atol::Float64 = 1.0e-08
)

Integrates an instance of SymbolicLinearKramersModel.

Important: inputs must be provided in international system (SI) units as a better physical practice. The only exception is the rotation rate ω provided in revolution multiples. If the discharge end is held by a dam, its height must be provided instead of the particle size, as it is used as the ODE initial condition.

model: a symbolic kiln model.

L: kiln length, [m].

R: kiln internal radius, [m].

Φ: kiln feed rate, [m³/s].

ω: kiln rotation rate, [rev/s].

β: kiln slope, [rad].

γ: solids repose angle, [rad].

d: particle size or dam height, [m].

1Kramers et al., 1952